+++ packages/terra-button-group/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated wdio screenshots due to functional-testing upgrade.
+  * Updated jest snapshots for icon button changes.
 
 ## 3.61.0 - (December 7, 2022)