+++ packages/terra-button/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Changed
   * Updated wdio screenshots due to functional-testing upgrade.
+* Added
+  * Added support for decorative icon buttons.
 
 ## 3.62.0 - (December 7, 2022)