+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Changed
   * Updated radio and checkbox tests to fix fusion theme tests.
+* Added
+  * Added A11y guide and examples for `terra-button`.
 
 ## 1.20.0 - (February 7, 2023)