+++ packages/terra-status-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated jest snapshots for icon button changes.
+
 ## 4.51.0 - (February 7, 2023)
 
 * Changed